Post Loading and Like buttons are implemented using Ajax

+++ senpaiGen/like/views.py
@@ -1,26 +1,34 @@
 from django.core.exceptions import ObjectDoesNotExist
 from .models import Like
 from post.models import Post
-from django.http import HttpResponseRedirect
+from django.http import JsonResponse
 
 
-def add_like(request, pk):
+def add_like(request):
 
-    post_ = Post.objects.get(pk=pk)
-    # a = Like.objects.filter(liked_user=request.user) & Like.objects.filter(post=post_)
+    if request.is_ajax():
 
-    try:
-        obj_ = Like.objects.get(liked_user=request.user, post=post_)  # check if already lik exist from current user or not .
-    except ObjectDoesNotExist:
-        obj_ = None
+        pk = request.POST.get('pk')
+        if pk is None:
+            return JsonResponse({'error': 'Invalid Post'})
 
-    if obj_:  # if the like from the current user exist on the same post
-        obj_.delete()
-        post_.total_no_of_likes -= 1
-        post_.save()
-        return HttpResponseRedirect("/")
+        post_ = Post.objects.get(pk=pk)
+        # a = Like.objects.filter(liked_user=request.user) & Like.objects.filter(post=post_)
 
-    else:
-        obj = Like(liked_user=request.user, post=post_, like=1)
-        obj.save()
-        return HttpResponseRedirect("/")
+        try:
+            obj_ = Like.objects.get(liked_user=request.user, post=post_)  # check if already lik exist from current user or not .
+        except ObjectDoesNotExist:
+            obj_ = None
+
+        if obj_:  # if the like from the current user exist on the same post
+            obj_.delete()
+            post_.total_no_of_likes -= 1
+            post_.save()
+            # return HttpResponseRedirect("/")
+            return JsonResponse({'status':200, 'total_no_of_likes':post_.total_no_of_likes, 'msg':'Like'})
+
+        else:
+            obj = Like(liked_user=request.user, post=post_, like=1)
+            obj.save()
+            # return HttpResponseRedirect("/")
+            return JsonResponse({'status':200, 'total_no_of_likes':post_.total_no_of_likes+1, 'msg':'Unlike'})
